hierarchy.jl

In these file there are useful functions to explore the potree, starting from the files continuing with the construction of the trie and, last but not least, the creation of a sub trie.

Main Interface

Potree.potree2trie — Function

potree2trie(potree::String) -> trie.children

Trie data structures for Potree hierarchy:

each file in octree directory is stored in a node,

first node is r, root of potree.

Potree.max_depth — Function

max_depth(trie::DataStructures.Trie{String}) -> Int

Maximum depth of trie.

Potree.cut_trie! — Function

cut_trie!(trie::DataStructures.Trie{String}, LOD::Int, l = 0::Int)

Return truncated trie.

Potree.get_all_values — Function

get_all_values(trie::DataStructures.Trie{String}) -> Array{String,1}

Return a collection of all values in trie.

Potree.sub_trie — Function

sub_trie(t::DataStructures.Trie{String}, root::AbstractString) -> DataStructures.Trie

Return the subtrie with defined root node.

Potree.get_files_in_potree_folder — Function

getfilesinpotreefolder(potree::String, LOD::Int, all_prev=true::Bool) -> Array{String,1}

Return all files at that level of potree

Potree.get_files — Function

getfiles(trie::DataStructures.Trie{String}, LOD::Int, data=String[]::Array{String,1}, l = 0::Int, allprev = true::Bool) -> Array{String,1}

Accumulate all values from root to defined level LOD.

Potree.readhrc — Function

readhrc(potree::String)

Reads the hrc file of the potree