utilities.jl

Some simple utilities functions similar to the ones present in json.jl and hierarchy.jl.

Main Interface

Potree.get_directories — Function

Read file line by line.

Potree.boxmodel_from_json — Function

boxmodelfromjson(volume::String)

Read volume model from a file JSON.

Potree.searchfile — Function

In recursive mode, search all files with key in filename.

Potree.source2pc — Function

generate input point cloud